refactor download code to separate file
